# docs/agentos-v2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DepDek AgentOS V2 design API",
    "version": "2.0.0-draft",
    "description": "Proposed P0 public transport boundary, NOT implemented. Business commands are versioned registered Manifests. Desktop v1 contract remains docs/contract.md."
  },
  "servers": [
    {
      "url": "https://depdek.local",
      "description": "Appliance TLS endpoint (placeholder, not an active service)"
    }
  ],
  "paths": {
...
    "/api/v2/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Authenticate final user",
        "description": "TLS and Origin validation required. Credentials must not be logged.",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful structured result.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "api_version": {
                      "type": "string",
                      "const": "2.0"
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "trace_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "workspace_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "policy_revision": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "data": {
                      "$ref": "#/components/schemas/Session"
                    }
                  },
                  "required": [
                    "api_version",
                    "request_id",
                    "trace_id",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "HttpOnly; Secure outside loopback development; SameSite=Strict.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/session": {
      "get": {
        "operationId": "session",
        "summary": "Current authenticated identity",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful structured result.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "api_version": {
                      "type": "string",
                      "const": "2.0"
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "trace_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "workspace_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "policy_revision": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "data": {
                      "$ref": "#/components/schemas/Session"
                    }
                  },
                  "required": [
                    "api_version",
                    "request_id",
                    "trace_id",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v2/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "Revoke this session",
        "security": [
          {
            "SessionCookie": [],
            "Csrf": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful structured result.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "api_version": {
                      "type": "string",
                      "const": "2.0"
                    },
                    "request_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "trace_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "workspace_id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "policy_revision": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "data": {
                      "$ref": "#/components/schemas/ReceiptSummary"
                    }
                  },
                  "required": [
                    "api_version",
                    "request_id",
                    "trace_id",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "depdek_session"
      },
      "Csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "api_version": {
            "type": "string",
            "const": "2.0"
          },
          "request_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "trace_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "required": [
          "api_version",
          "request_id",
          "trace_id",
          "error"
        ],
        "additionalProperties": false
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1024,
            "writeOnly": true
          }
        },
        "required": [
          "username",
          "password"
        ],
        "additionalProperties": false
      },
      "Session": {
        "type": "object",
        "properties": {
          "principal_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "display_name": {
            "type": "string"
          },
          "authentication_epoch": {
            "type": "integer",
            "minimum": 1
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "csrf_token": {
            "type": "string"
          },
          "workspaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Workspace"
            }
          }
        },
        "required": [
          "principal_id",
          "authentication_epoch",
          "expires_at",
          "csrf_token",
          "workspaces"
        ],
        "additionalProperties": false
      },
      "Workspace": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "kind": {
            "type": "string",
            "enum": [
              "personal",
              "family"
            ]
          },
          "display_name": {
            "type": "string"
          },
          "policy_revision": {
            "type": "integer",
            "minimum": 1
          },
          "locked": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "kind",
          "policy_revision",
          "locked"
        ],
        "additionalProperties": false
      },
...
      "ReceiptSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "outcome": {
            "type": "string"
          },
          "verified": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "outcome",
          "verified"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "Error": {
        "description": "Stable error envelope; see api.md for code and HTTP status mapping.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  },
...
}